feat: extract event logic for select component
